refactor(auth): remove deprecated controllers and implement new registration flow
- Deleted obsolete controllers: CompleteRegisterController, ForgotPasswordController, LoginController, PreRegisterController, RegisterController, ResetPasswordController.
- Updated auth routes to reflect the removal of these controllers.
- Introduced a new CompleteRegisterController to handle the complete registration process with email verification.
- Refactored the Auth\Controller to centralize common functionalities for authentication-related controllers.
