+++ src/Http/Controllers/Auth/RegisterController.php
@@ -1,0 +1,145 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Unusualify\Modularity\Http\Controllers\Auth;
+
+use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Route;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Validation\Rules;
+use Unusualify\Modularity\Entities\Company;
+use Unusualify\Modularity\Events\ModularityUserRegistered;
+use Unusualify\Modularity\Events\ModularityUserRegistering;
+use Unusualify\Modularity\Services\MessageStage;
+
+class RegisterController extends Controller
+{
+    public function showForm()
+    {
+        if (modularityConfig('email_verified_register')) {
+            return redirect()->route(Route::hasAdmin('register.email_form'));
+        }
+
+        return $this->viewFactory->make(modularityBaseKey() . '::auth.register', [
+            'attributes' => $this->authBannerAttributes(),
+            'formAttributes' => array_merge(
+                ['title' => $this->authFormTitle(__('authentication.create-an-account'), ['transform' => ''])],
+                $this->authFormBaseAttributes(
+                    'register_form',
+                    route(Route::hasAdmin('register')),
+                    'authentication.register'
+                )
+            ),
+            'formSlots' => $this->haveAccountOptionSlot(),
+            'slots' => [
+                'bottom' => $this->authBottomSlots([
+                    $this->oauthGoogleButtonSlot('sign-up'),
+                ]),
+            ],
+        ]);
+    }
+
+    /**
+     * Get a validator for an incoming registration request.
+     *
+     * @return \Illuminate\Contracts\Validation\Validator
+     */
+    protected function validator(array $data)
+    {
+        return Validator::make($data, $this->rules());
+    }
+
+    /**
+     * Create a new user instance after a valid registration.
+     *
+     * @param array $data
+     * @return \App\Models\User
+     */
+    protected function register(Request $request)
+    {
+        $emailVerifiedRegister = modularityConfig('email_verified_register');
+
+        if ($emailVerifiedRegister) {
+            return $request->wantsJson()
+                ? new JsonResponse([
+                    'variant' => MessageStage::ERROR,
+                    'message' => 'Restricted Registration',
+                    'redirector' => route(Route::hasAdmin('register.email_form')),
+                    'login_page' => route(Route::hasAdmin('login.form')),
+                ], 200)
+                : redirect()->route(Route::hasAdmin('register.email_form'));
+        }
+
+        $validator = $this->validator($request->all());
+
+        if ($validator->fails()) {
+            return $request->wantsJson()
+                ? new JsonResponse([
+                    'errors' => $validator->errors(),
+                    'message' => $validator->messages()->first(),
+                    'variant' => MessageStage::WARNING,
+                ], 422)
+                : $request->validate($this->rules());
+
+            return $res;
+        }
+
+        event(new ModularityUserRegistering($request));
+
+        $user = Company::create([
+            'name' => $request['company'] ?? '',
+            'spread_payload' => [
+                'is_personal' => $request['company'] ? false : true,
+            ],
+        ])->users()->create([
+            'name' => $request['name'],
+            'email' => $request['email'],
+            'password' => Hash::make($request['password']),
+            'language' => $request['language'] ?? app()->getLocale(),
+        ]);
+
+        $user->assignRole('client-manager');
+
+        event(new ModularityUserRegistered($user, $request));
+
+        return $request->wantsJson()
+            ? new JsonResponse([
+                'status' => 'success',
+                'message' => 'User registered successfully',
+                'redirector' => route(Route::hasAdmin('register.success')),
+                'login_page' => route(Route::hasAdmin('login')),
+            ], 200)
+            : $this->sendLoginResponse($request);
+    }
+
+    public function rules()
+    {
+        $usersTable = modularityConfig('tables.users', 'um_users');
+
+        return [
+            'name' => ['required', 'string', 'max:255'],
+            // Surname is not mandatory.
+            'surname' => ['required', 'string', 'max:255'],
+            // 'company' => ['required', 'string', 'max:255'],
+            'email' => ['required', 'string', 'email', 'max:255', 'unique:' . $usersTable . ',email'],
+            'password' => ['required', 'confirmed', Rules\Password::defaults()],
+            // 'tos' => ['required', 'boolean'],
+        ];
+    }
+
+    public function success()
+    {
+        return view(modularityBaseKey() . '::auth.success', [
+            'taskState' => [
+                'status' => 'success',
+                'title' => __('authentication.register-title'),
+                'description' => __('authentication.register-description'),
+                'button_text' => __('authentication.register-button-text'),
+                'button_url' => route('admin.login'),
+            ],
+        ]);
+    }
+}
